+++ examples/live/bitmex/bitmex_data_tester.py
@@ -38,6 +38,7 @@ from nautilus_trader.test_kit.strategies.tester_data import DataTesterConfig
 # Futures: XBTH25 (Bitcoin futures expiring March 2025)
 # Alt perpetuals: ETHUSD, SOLUSD, etc.
 
+testnet = False  # If client uses the testnet API
 symbol = "XBTUSD"  # Bitcoin perpetual swap
 # symbol = "SOLUSDT"  # Solana spot
 # symbol = "ETHUSDT"  # Ethereum spot
@@ -57,12 +58,12 @@ config_node = TradingNodeConfig(
             base_url_http=None,  # Override with custom endpoint
             base_url_ws=None,  # Override with custom endpoint
             instrument_provider=InstrumentProviderConfig(load_all=True),
-            testnet=False,  # If client uses the testnet API
+            testnet=testnet,  # If client uses the testnet API
         ),
     },
-    timeout_connection=30.0,
+    timeout_connection=10.0,
     timeout_reconciliation=10.0,
-    timeout_disconnection=10.0,
+    timeout_disconnection=2.0,
     timeout_post_stop=1.0,
 )