+++ nautilus_trader/adapters/bitmex/data.py
@@ -129,6 +129,7 @@ class BitmexDataClient(LiveMarketDataClient):
             api_secret=config.api_secret,
             account_id=None,  # Not required for data
             heartbeat=30,
+            testnet=config.testnet,
         )
         self._ws_client_futures: set[asyncio.Future] = set()
         self._log.info(f"WebSocket URL {ws_url}", LogColor.BLUE)